Split eventHandler into separate handlers per main feature
