+++ ios/THEOplayerRCTView.swift
@@ -7,7 +7,9 @@ import THEOplayerSDK
 class THEOplayerRCTView: UIView {
     // MARK: Members
     private var player: THEOplayer?
-    private var eventHandler: THEOplayerRCTViewEventHandler
+    private var mainEventHandler: THEOplayerRCTViewMainEventHandler
+    private var textTrackEventHandler: THEOplayerRCTViewTextTrackEventHandler
+    private var adEventHandler: THEOplayerRCTViewAdEventHandler
     
     // MARK: Bridged props
     private var src: SourceDescription?
@@ -29,14 +31,18 @@ class THEOplayerRCTView: UIView {
     
     // MARK: - Initialisation / view setup
     init() {
-        // create an event handler to contain event props
-        self.eventHandler = THEOplayerRCTViewEventHandler()
+        // create event handlers to maintain event props
+        self.mainEventHandler = THEOplayerRCTViewMainEventHandler()
+        self.textTrackEventHandler = THEOplayerRCTViewTextTrackEventHandler()
+        self.adEventHandler = THEOplayerRCTViewAdEventHandler()
         
         super.init(frame: .zero)
     }
     
     func destroy() {
-        self.eventHandler.destroy()
+        self.mainEventHandler.destroy()
+        self.textTrackEventHandler.destroy()
+        self.adEventHandler.destroy()
         self.player?.destroy()
         self.player = nil
         if DEBUG_THEOPLAYER_INTERACTION { print("[NATIVE] THEOplayer instance destroyed.") }
@@ -88,8 +94,10 @@ class THEOplayerRCTView: UIView {
                 // Therefore, dispatched to next runloop tick
                 self.initPlayer()
                 if let player = self.player {
-                    // couple player instance to event handler
-                    self.eventHandler.setPlayer(player)
+                    // couple player instance to event handlers
+                    self.mainEventHandler.setPlayer(player)
+                    self.textTrackEventHandler.setPlayer(player)
+                    self.adEventHandler.setPlayer(player)
                     // couple player instance to view
                     player.addAsSubview(of: self)
                 }
@@ -332,130 +340,136 @@ class THEOplayerRCTView: UIView {
         }
     }
     
-    // MARK: - Listener based event bridging
+    // MARK: - Listener based MAIN event bridging
+    
     @objc(setOnNativePlay:)
     func setOnNativePlay(nativePlay: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativePlay = nativePlay
+        self.mainEventHandler.onNativePlay = nativePlay
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativePlay prop set.") }
     }
     
     @objc(setOnNativePause:)
     func setOnNativePause(nativePause: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativePause = nativePause
+        self.mainEventHandler.onNativePause = nativePause
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativePause prop set.") }
     }
     
     @objc(setOnNativeSourceChange:)
     func setOnNativeSourceChange(nativeSourceChange: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeSourceChange = nativeSourceChange
+        self.mainEventHandler.onNativeSourceChange = nativeSourceChange
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeSourceChange prop set.") }
     }
     
     @objc(setOnNativeLoadStart:)
     func setOnNativeLoadStart(nativeLoadStart: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeLoadStart = nativeLoadStart
+        self.mainEventHandler.onNativeLoadStart = nativeLoadStart
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeLoadStart prop set.") }
     }
     
     @objc(setOnNativeReadyStateChange:)
     func setOnNativeReadyStateChange(nativeReadyStateChange: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeReadyStateChange = nativeReadyStateChange
+        self.mainEventHandler.onNativeReadyStateChange = nativeReadyStateChange
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeReadyStateChange prop set.") }
     }
     
     @objc(setOnNativeDurationChange:)
     func setOnNativeDurationChange(nativeDurationChange: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeDurationChange = nativeDurationChange
+        self.mainEventHandler.onNativeDurationChange = nativeDurationChange
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeDurationChange prop set.") }
     }
     
     @objc(setOnNativeProgress:)
     func setOnNativeProgress(nativeProgress: @escaping RCTBubblingEventBlock) {
-        self.eventHandler.onNativeProgress = nativeProgress
+        self.mainEventHandler.onNativeProgress = nativeProgress
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeProgress prop set.") }
     }
     
     @objc(setOnNativeTimeUpdate:)
     func setOnNativeTimeUpdate(nativeTimeUpdate: @escaping RCTBubblingEventBlock) {
-        self.eventHandler.onNativeTimeUpdate = nativeTimeUpdate
+        self.mainEventHandler.onNativeTimeUpdate = nativeTimeUpdate
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeTimeUpdate prop set.") }
     }
     
     @objc(setOnNativePlaying:)
     func setOnNativePlaying(nativePlaying: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativePlaying = nativePlaying
+        self.mainEventHandler.onNativePlaying = nativePlaying
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativePlaying prop set.") }
     }
     
     @objc(setOnNativeSeeking:)
     func setOnNativeSeeking(nativeSeeking: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeSeeking = nativeSeeking
+        self.mainEventHandler.onNativeSeeking = nativeSeeking
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeSeeking prop set.") }
     }
     
     @objc(setOnNativeSeeked:)
     func setOnNativeSeeked(nativeSeeked: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeSeeked = nativeSeeked
+        self.mainEventHandler.onNativeSeeked = nativeSeeked
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeSeeked prop set.") }
     }
     
     @objc(setOnNativeEnded:)
     func setOnNativeEnded(nativeEnded: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeEnded = nativeEnded
+        self.mainEventHandler.onNativeEnded = nativeEnded
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeEnded prop set.") }
     }
     
     @objc(setOnNativeError:)
     func setOnNativeError(nativeError: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeError = nativeError
+        self.mainEventHandler.onNativeError = nativeError
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeError prop set.") }
     }
     
     @objc(setOnNativeLoadedData:)
     func setOnNativeLoadedData(nativeLoadedData: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeLoadedData = nativeLoadedData
+        self.mainEventHandler.onNativeLoadedData = nativeLoadedData
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeLoadedData prop set.") }
     }
     
     @objc(setOnNativeLoadedMetadata:)
     func setOnNativeLoadedMetadata(nativeLoadedMetadata: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeLoadedMetadata = nativeLoadedMetadata
+        self.mainEventHandler.onNativeLoadedMetadata = nativeLoadedMetadata
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeLoadedMetadata prop set.") }
     }
     
     @objc(setOnNativeFullscreenPlayerWillPresent:)
     func setOnNativeFullscreenPlayerWillPresent(nativeFullscreenPlayerWillPresent: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeFullscreenPlayerWillPresent = nativeFullscreenPlayerWillPresent
+        self.mainEventHandler.onNativeFullscreenPlayerWillPresent = nativeFullscreenPlayerWillPresent
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeFullscreenPlayerWillPresent prop set.") }
     }
     
     @objc(setOnNativeFullscreenPlayerDidPresent:)
     func setOnNativeFullscreenPlayerDidPresent(nativeFullscreenPlayerDidPresent: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeFullscreenPlayerDidPresent = nativeFullscreenPlayerDidPresent
+        self.mainEventHandler.onNativeFullscreenPlayerDidPresent = nativeFullscreenPlayerDidPresent
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeFullscreenPlayerDidPresent prop set.") }
     }
 
     @objc(setOnNativeFullscreenPlayerWillDismiss:)
     func setOnNativeFullscreenPlayerWillDismiss(nativeFullscreenPlayerWillDismiss: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeFullscreenPlayerWillDismiss = nativeFullscreenPlayerWillDismiss
+        self.mainEventHandler.onNativeFullscreenPlayerWillDismiss = nativeFullscreenPlayerWillDismiss
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeFullscreenPlayerWillDismiss prop set.") }
     }
     
     @objc(setOnNativeFullscreenPlayerDidDismiss:)
     func setOnNativeFullscreenPlayerDidDismiss(nativeFullscreenPlayerDidDismiss: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeFullscreenPlayerDidDismiss = nativeFullscreenPlayerDidDismiss
+        self.mainEventHandler.onNativeFullscreenPlayerDidDismiss = nativeFullscreenPlayerDidDismiss
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeFullscreenPlayerDidDismiss prop set.") }
     }
     
+    // MARK: - Listener based TEXTTRACK event bridging
+    
     @objc(setOnNativeTextTrackListEvent:)
     func setOnNativeTextTrackListEvent(nativeTextTrackListEvent: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeTextTrackListEvent = nativeTextTrackListEvent
+        self.textTrackEventHandler.onNativeTextTrackListEvent = nativeTextTrackListEvent
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeTextTrackListEvent prop set.") }
     }
     
     @objc(setOnNativeTextTrackEvent:)
     func setOnNativeTextTrackEvent(nativeTextTrackEvent: @escaping RCTDirectEventBlock) {
-        self.eventHandler.onNativeTextTrackEvent = nativeTextTrackEvent
+        self.textTrackEventHandler.onNativeTextTrackEvent = nativeTextTrackEvent
         if DEBUG_PROP_UPDATES  { print("[NATIVE] nativeTextTrackEvent prop set.") }
     }
+    
+    // MARK: - Listener based AD event bridging
+    // todo
 }
