+++ ios/THEOplayerRCTViewMainEventHandler.swift
@@ -3,13 +3,7 @@
 import Foundation
 import THEOplayerSDK
 
-let ADD_TRACK: Int = 0
-let REMOVE_TRACK: Int = 1
-let ADD_CUE: Int = 0
-let REMOVE_CUE: Int = 1
-
-
-class THEOplayerRCTViewEventHandler {
+class THEOplayerRCTViewMainEventHandler {
     // MARK: Members
     private weak var player: THEOplayer?
     private var currentPresentationMode = THEOplayerSDK.PresentationMode.inline // TheoPlayer's initial presentationMode
@@ -34,8 +28,6 @@ class THEOplayerRCTViewEventHandler {
     var onNativeFullscreenPlayerDidPresent: RCTDirectEventBlock?
     var onNativeFullscreenPlayerWillDismiss: RCTDirectEventBlock?
     var onNativeFullscreenPlayerDidDismiss: RCTDirectEventBlock?
-    var onNativeTextTrackListEvent: RCTDirectEventBlock?
-    var onNativeTextTrackEvent: RCTDirectEventBlock?
     
     // MARK: player Listeners
     private var playListener: EventListener?
@@ -55,15 +47,6 @@ class THEOplayerRCTViewEventHandler {
     private var loadedMetadataListener: EventListener?
     private var presentationModeChangeListener: EventListener?
     
-    // MARK: textTrackList Listeners
-    private var addTrackListener: EventListener?
-    private var removeTrackListener: EventListener?
-    
-    // MARK: textTrack listeners (attached dynamically to new texttracks)
-    private var addCueListeners: [Int:EventListener] = [:]
-    private var removeCueListeners: [Int:EventListener] = [:]
-    
-    
     // MARK: - destruction
     func destroy() {
         // dettach listeners
@@ -79,17 +62,15 @@ class THEOplayerRCTViewEventHandler {
     }
     
     private func attachListeners() {
-        self.attachPlayerListeners()
-        self.attachTextTrackListListeners()
+        self.attachMainPlayerListeners()
     }
     
     private func dettachListeners() {
-        self.dettachPlayerListeners()
-        self.dettachTextTrackListListeners()
+        self.dettachMainPlayerListeners()
     }
     
-    // MARK: - attach/dettach player Listeners
-    private func attachPlayerListeners() {
+    // MARK: - attach/dettach main player Listeners
+    private func attachMainPlayerListeners() {
         guard let player = self.player else {
             return
         }
@@ -296,7 +277,7 @@ class THEOplayerRCTViewEventHandler {
         if DEBUG_EVENTHANDLER { print("[NATIVE] PresentationModeChange listener attached to THEOplayer") }
     }
     
-    private func dettachPlayerListeners() {
+    private func dettachMainPlayerListeners() {
         guard let player = self.player else {
             return
         }
@@ -395,99 +376,6 @@ class THEOplayerRCTViewEventHandler {
         if let presentationModeChangeListener = self.presentationModeChangeListener {
             player.removeEventListener(type: PlayerEventTypes.PRESENTATION_MODE_CHANGE, listener: presentationModeChangeListener)
             if DEBUG_EVENTHANDLER { print("[NATIVE] PresentationModeChange listener dettached from THEOplayer") }
-        }
-    }
-    
-    // MARK: - attach/dettach textTrackList Listeners
-    private func attachTextTrackListListeners() {
-        guard let player = self.player else {
-            return
-        }
-        
-        // ADD_TRACK
-        self.addTrackListener = player.textTracks.addEventListener(type: TextTrackListEventTypes.ADD_TRACK) { [weak self] event in
-            guard let welf = self else { return }
-            if let forwardedTextTrackListEvent = welf.onNativeTextTrackListEvent,
-               let textTrack = event.track as? TextTrack {
-                if DEBUG_THEOPLAYER_EVENTS { print("[NATIVE] Received ADD_TRACK event from THEOplayer textTrack list: trackUid = \(textTrack.uid)") }
-                // trigger tracklist event
-                forwardedTextTrackListEvent([
-                    "track" : THEOplayerRCTMetadataAggregator.aggregatedTextTrackInfo(textTrack: textTrack),
-                    "type" : ADD_TRACK
-                ])
-                // start listening for cue events on this track and keep listener for later removal
-                welf.addCueListeners[textTrack.uid] = textTrack.addEventListener(type: TextTrackEventTypes.ADD_CUE, listener: welf.addCueListener(_:))
-                if DEBUG_EVENTHANDLER { print("[NATIVE] AddCue listener attached to THEOplayer textTrack with uid \(textTrack.uid)") }
-                welf.removeCueListeners[textTrack.uid] = textTrack.addEventListener(type: TextTrackEventTypes.REMOVE_CUE, listener: welf.removeCueListener(_:))
-                if DEBUG_EVENTHANDLER { print("[NATIVE] RemoveCue listener attached to THEOplayer textTrack with uid \(textTrack.uid)") }
-            }
-        }
-        if DEBUG_EVENTHANDLER { print("[NATIVE] AddTrack listener attached to THEOplayer textTrack list") }
-        
-        // REMOVE_TRACK
-        self.removeTrackListener = player.textTracks.addEventListener(type: TextTrackListEventTypes.REMOVE_TRACK) { [weak self] event in
-            guard let welf = self else { return }
-            if let forwardedTextTrackListEvent = welf.onNativeTextTrackListEvent,
-               let textTrack = event.track as? TextTrack {
-                if DEBUG_THEOPLAYER_EVENTS { print("[NATIVE] Received REMOVE_TRACK event from THEOplayer textTrack list: trackUid = \(textTrack.uid)") }
-                // trigger tracklist event
-                forwardedTextTrackListEvent([
-                    "track" : THEOplayerRCTMetadataAggregator.aggregatedTextTrackInfo(textTrack: textTrack),
-                    "type" : REMOVE_TRACK
-                ])
-                // stop listening for cue events on this track
-                if let addCueListener = welf.addCueListeners[textTrack.uid],
-                   let removeCueListener = welf.removeCueListeners[textTrack.uid]{
-                    textTrack.removeEventListener(type: TextTrackEventTypes.ADD_CUE, listener: addCueListener)
-                    if DEBUG_EVENTHANDLER { print("[NATIVE] AddCue listener removed from THEOplayer textTrack with uid \(textTrack.uid)") }
-                    textTrack.removeEventListener(type: TextTrackEventTypes.REMOVE_CUE, listener: removeCueListener)
-                    if DEBUG_EVENTHANDLER { print("[NATIVE] RemoveCue listener removed from THEOplayer textTrack with uid \(textTrack.uid)") }
-                }
-            }
-        }
-        if DEBUG_EVENTHANDLER { print("[NATIVE] RemoveTrack listener attached to THEOplayer textTrack list") }
-    }
-    
-    private func dettachTextTrackListListeners() {
-        guard let player = self.player else {
-            return
-        }
-        
-        // ADD_TRACK
-        if let addTrackListener = self.addTrackListener {
-            player.textTracks.removeEventListener(type: TextTrackListEventTypes.ADD_TRACK, listener: addTrackListener)
-            if DEBUG_EVENTHANDLER { print("[NATIVE] AddTrack listener dettached from THEOplayer textTrack list") }
-        }
-        
-        // REMOVE_TRACK
-        if let removeTrackListener = self.removeTrackListener {
-            player.textTracks.removeEventListener(type: TextTrackListEventTypes.REMOVE_TRACK, listener: removeTrackListener)
-            if DEBUG_EVENTHANDLER { print("[NATIVE] RemoveTrack listener dettached from THEOplayer textTrack list") }
-        }
-    }
-    
-    // MARK: - dynamic textTrack Listeners
-    private func addCueListener(_ event: AddCueEvent) {
-        if let forwardedTextTrackEvent = self.onNativeTextTrackEvent,
-           let textTrack = event.cue.track {
-            if DEBUG_THEOPLAYER_EVENTS { print("[NATIVE] Received ADD_CUE event from textTrack: trackUid = \(textTrack.uid), cueUid = \(event.cue.uid)") }
-            forwardedTextTrackEvent([
-                "trackUid" : textTrack.uid,
-                "type": ADD_CUE,
-                "cue": THEOplayerRCTMetadataAggregator.aggregatedTextTrackCueInfo(textTrackCue: event.cue)
-            ])
-        }
-    }
-    
-    private func removeCueListener(_ event: RemoveCueEvent) {
-        if let forwardedTextTrackEvent = self.onNativeTextTrackEvent,
-           let textTrack = event.cue.track {
-            if DEBUG_THEOPLAYER_EVENTS { print("[NATIVE] Received REMOVE_CUE event from textTrack: trackUid = \(textTrack.uid), cueUid = \(event.cue.uid)") }
-            forwardedTextTrackEvent([
-                "trackUid" : textTrack.uid,
-                "type": REMOVE_CUE,
-                "cue": THEOplayerRCTMetadataAggregator.aggregatedTextTrackCueInfo(textTrackCue: event.cue)
-            ])
         }
     }
 }
